ui: Improvements in joystick handling
- Use of SDL game controller subsystem instead of the raw joystick
  interface; this way macOS and Linux share the same code.

- Added buttons: A, B, X, Y, Guide, Start, Back to the emulated
  joystick.
  Not used by any of the current emulated platforms, though.

- Prepend global namespace to SDL methods.

+++ src/core/kempston.hpp
@@ -32,12 +32,12 @@ namespace kempston {
  * |       |       |       | Fire  | Up    | Down  | Left  | Right |
  * +-------+-------+-------+-------+-------+-------+-------+-------+
  */
-constexpr static const uint8_t JOY_PORT_UP    = D3;
-constexpr static const uint8_t JOY_PORT_DOWN  = D2;
-constexpr static const uint8_t JOY_PORT_LEFT  = D1;
-constexpr static const uint8_t JOY_PORT_RIGHT = D0;
-constexpr static const uint8_t JOY_PORT_FIRE  = D4;
-constexpr static const uint8_t JOY_MASK       = D0 | D1 | D2 | D3 | D4;
+constexpr static const uint16_t JOY_PORT_UP    = D3;
+constexpr static const uint16_t JOY_PORT_DOWN  = D2;
+constexpr static const uint16_t JOY_PORT_LEFT  = D1;
+constexpr static const uint16_t JOY_PORT_RIGHT = D0;
+constexpr static const uint16_t JOY_PORT_FIRE  = D4;
+constexpr static const uint16_t JOY_MASK       = D0 | D1 | D2 | D3 | D4;
 
 /**
  * Kempston interface joystick port.
